feat: migrate to Cloudflare Pages (SSR) with native /api/popularity endpoint
