--- packages/popularity-worker/package.json
@@ -1,13 +1,0 @@
-{
-	"name": "bansos-popularity-worker",
-	"version": "0.0.1",
-	"private": true,
-	"type": "module",
-	"scripts": {
-		"deploy": "wrangler deploy",
-		"dev": "wrangler dev"
-	},
-	"devDependencies": {
-		"wrangler": "^3.109.0"
-	}
-}
