--- packages/popularity-worker/wrangler.toml
@@ -1,10 +1,0 @@
-name = "bansos-popularity-worker"
-main = "src/index.ts"
-compatibility_date = "2026-06-18"
-
-[[routes]]
-pattern = "api.bansos.dev/*"
-custom_domain = true
-
-[observability]
-enabled = true
