+++ src/app.d.ts
@@ -13,7 +13,12 @@ declare global {
 		// interface Locals {}
 		// interface PageData {}
 		// interface PageState {}
-		// interface Platform {}
+		interface Platform {
+			env: {
+				CF_API_TOKEN?: string;
+				CF_ZONE_ID?: string;
+			};
+		}
 	}
 }
